Add parser examples and AST parser tests for multiple languages
